Move helper from sites config to multitenant config
